Move and simplify global client telemetry test
